Update README.md
Please read through and follow the readme carefully to run the application, and i did not create menus for the routes. You must create a user and promte the user as admin through console as how it is mentioned and follow the app links to play around the functionalities required. Only user with admin permission can access the app fucntionality links.